requirements.txt
mock
nose
pika
redis
PyYAML
protobuf
flask
python-consul
prometheus_client>=0.0.18
requests
biomaj_core
biomaj_zipkin
